Refactored Fabric Interop CC libs test to ensure compilation
