+++ core/network/fabric-interop-cc/libs/test/go.mod
@@ -4,9 +4,12 @@ go 1.16
 
 require (
 	github.com/golang/protobuf v1.5.2
-	github.com/hyperledger-labs/weaver-dlt-interoperability/core/network/fabric-interop-cc/contracts/interop v1.2.2
 	github.com/hyperledger/fabric-chaincode-go v0.0.0-20210718160520-38d29fabecb9
 	github.com/hyperledger/fabric-contract-api-go v1.1.1
 	github.com/hyperledger/fabric-protos-go v0.0.0-20210720123151-f0dc3e2a0871
 	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.6.1 // indirect
+	google.golang.org/grpc v1.39.0 // indirect
+	google.golang.org/protobuf v1.27.1 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
 )
