+++ core/network/fabric-interop-cc/libs/test/setup_test.go
@@ -7,61 +7,17 @@
 // test contains the stubs required for counterfeiter to generate the mocks used for tests
 //
 // The mocks are created by running `go generate ./...` as described in the counterfeiter readme
-package main
+package test
 
 import (
 	"os"
 	"testing"
 
-	"github.com/hyperledger-labs/weaver-dlt-interoperability/core/network/fabric-interop-cc/contracts/interop/mocks"
-	"github.com/hyperledger/fabric-chaincode-go/pkg/cid"
-	"github.com/hyperledger/fabric-chaincode-go/shim"
-	"github.com/hyperledger/fabric-contract-api-go/contractapi"
 	log "github.com/sirupsen/logrus"
-)
-
-const (
-	myOrg1Msp      = "Org1Testmsp"
-	myOrg1Clientid = "myOrg1Userid"
 )
 
 func TestMain(m *testing.M) {
 	log.SetLevel(log.PanicLevel)
 	log.SetOutput(os.Stdout)
 	os.Exit(m.Run())
-}
-
-//go:generate counterfeiter -o mocks/transaction.go -fake-name TransactionContext . transactionContext
-type transactionContext interface {
-	contractapi.TransactionContextInterface
-}
-
-//go:generate counterfeiter -o mocks/chaincodestub.go -fake-name ChaincodeStub . chaincodeStub
-type chaincodeStub interface {
-	shim.ChaincodeStubInterface
-}
-
-//go:generate counterfeiter -o mocks/clientIdentity.go -fake-name ClientIdentity . clientIdentity
-type clientIdentity interface {
-	cid.ClientIdentity
-}
-
-func prepMockStub() (*mocks.TransactionContext, *mocks.ChaincodeStub, SmartContract) {
-	transactionContext, chaincodeStub := prepMocks(myOrg1Msp, myOrg1Clientid)
-	interopcc := SmartContract{}
-	return transactionContext, chaincodeStub, interopcc
-}
-
-func prepMocks(orgMSP, clientID string) (*mocks.TransactionContext, *mocks.ChaincodeStub) {
-	chaincodeStub := &mocks.ChaincodeStub{}
-	transactionContext := &mocks.TransactionContext{}
-	transactionContext.GetStubReturns(chaincodeStub)
-
-	clientIdentity := &mocks.ClientIdentity{}
-	clientIdentity.GetMSPIDReturns(orgMSP, nil)
-	clientIdentity.GetIDReturns(clientID, nil)
-	//set matching msp ID using peer shim env variable
-	os.Setenv("CORE_PEER_LOCALMSPID", orgMSP)
-	transactionContext.GetClientIdentityReturns(clientIdentity)
-	return transactionContext, chaincodeStub
 }
